+++ test/integration/default/serverspec/spec_helper.rb
@@ -2,4 +2,9 @@
 
 require 'serverspec'
 
-set :backend, :exec
+if RUBY_PLATFORM.match(/mswin|mingw32|windows/)
+  set :os, family: 'windows'
+  set :backend, :cmd
+else
+  set :backend, :exec
+end