+++ metadata.rb
@@ -7,7 +7,7 @@ maintainer_email 'j@p4nt5.com'
 license          'apache2'
 description      'Installs/Configures Divvy'
 long_description 'Installs/Configures Divvy'
-version          '0.1.1'
+version          '0.2.0'
 
 depends          'mac-app-store', '~> 1.0'
 depends          'macosx_accessibility', '~> 1.0'