+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Divvy Cookbook CHANGELOG
 
 v?.?.? (????-??-??)
 -------------------
+- Add support for non-App Store direct downloads for OS X
+- Add support for Windows
 
 v0.1.0 (2015-04-25)
 -------------------