+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 Divvy Cookbook CHANGELOG
 ========================
 
-v?.?.? (????-??-??)
+v0.2.0 (2015-05-09)
 -------------------
 - Add support for non-App Store direct downloads for OS X
 - Add support for Windows
 - Grant required Accessibility privileges automatically on OS X
-- Add a `:run` action and start Divvy after it's installed
+- Add a `:start` action and start Divvy after it's installed
 - Add an `:enable` action to start Divvy on log-in and do it after install
 
 v0.1.0 (2015-04-25)